Integrated slimlock into slim
iwamatsu's final release of SLiM (1.3.6) had integrated slimlock into
the same codebase.  This commit foward-ports the same code from the
1.3.6 release that was on berlios before berlios was discontinued.

Note there might be some other code changes that are not directly
slimlock-related in this commit; it was difficult to separate them out.

The slim-1.3.6 release had a number of inconsistencies related to the
slimlock merge that can cause build and runtime errors, so additionally
this commit addresses those:

- added missing link deps to libslim
- ensured slimlock isn't installed when it isn't built
- install slimlock SUID, as it was in its predecessor