Prevent PAM Session from dying in child proc
There is a call to pam.end() shortly after forking the child process, in
app.cpp.  This is erroneous as it kills the active pam session.  In a default
PAM setup this has no effect, but when PAM has been set up with kerberos
integration, the kerberos ticket cache is cleaned.  This may also have an effect
on SElinux systems.

This bug is the underlying reason why erroneous patches 2690 and (part of)
3043 [in the BerliOS system] have been submitted to attempt to fix kerberos
issues, however the actual fix is to use this patch.

https://bugs.gentoo.org/show_bug.cgi?id=436556