Add support for running X setup commands at slim startup
This patch adds support similar to that which can be achieved
in other DM's (such as XDM) by customizing or editing their
Xsetup scripts.  This was added in particular to support
nvidia-drivers based Optimus setups, which need to execute
two xrandr commands in order for Xorg to render on-screen
correctly.  Usage is simple: write a script with the needed
commands, ensure it is executable, and set the xsetup_script
option in slim.conf to the full path of this script.