app.cpp: Hard code mcookiesize
Under certain conditions, mcookiesize as it's used in the for loop
of App::CreateServerAuth() is not providing the usual 32 that is
hard-coded into the object creator, but rather a random large integer.
This in turn can cause segfaults.  Attempting to work around this through
using a #define to hard-code the length of mcookie instead of using the
const int variable.

https://bugs.gentoo.org/show_bug.cgi?id=608816